requirements.txt
pandas
numpy
keras
tensorflow
scikit-learn
pyaml
matplotlib
flask
pyserial